requirements.txt
fastapi
newspaper3k
uvicorn
lxml_html_clean